+++ pypdl/main.py
@@ -1,18 +1,7 @@
 import logging
-import time
-from concurrent.futures import ThreadPoolExecutor, as_completed, Future
-from typing import Callable, Optional, Union
-from threading import Event
 
-from manager import DownloadManager
-from utls import (
-    AutoShutdownFuture,
-    FileValidator,
-    ScreenCleaner,
-    cursor_up,
-    seconds_to_hms,
-    to_mb,
-)
+from .factory import Factory
+from .manager import DownloadManager
 
 handler = logging.FileHandler("pypdl.log", mode="a", delay=True)
 handler.setFormatter(
@@ -27,309 +16,13 @@ class Pypdl(DownloadManager):
     """
     A multi-threaded file downloader that supports progress tracking, retries, pause/resume functionality etc.
 
-    Keyword Arguments:
-        params (dict, optional): A dictionary, list of tuples or bytes to send as a query string. Default is None.
-        allow_redirects (bool, optional): A Boolean to enable/disable redirection. Default is True.
-        auth (tuple, optional): A tuple to enable a certain HTTP authentication. Default is None.
-        cert (str or tuple, optional): A String or Tuple specifying a cert file or key. Default is None.
-        cookies (dict, optional): A dictionary of cookies to send to the specified url. Default is None.
-        headers (dict, optional): A dictionary of HTTP headers to send to the specified url. Default is None.
-        proxies (dict, optional): A dictionary of the protocol to the proxy url. Default is None.
-        timeout (number or tuple, optional): A number, or a tuple, indicating how many seconds to wait for the client to make a connection and/or send a response. Default is 10 seconds.
-        verify (bool or str, optional): A Boolean or a String indication to verify the servers TLS certificate or not. Default is True.
+    This class accepts keyword arguments that are valid for aiohttp.ClientSession.
     """
 
-    def __init__(self, allow_reuse=False, **kwargs):
-        super().__init__(**kwargs)
-        self._allow_reuse = allow_reuse
 
-    def start(
-        self,
-        url: str,
-        file_path: Optional[str] = None,
-        segments: int = 10,
-        display: bool = True,
-        multisegment: bool = True,
-        block: bool = True,
-        retries: int = 0,
-        mirror_func: Optional[Callable[[], str]] = None,
-        etag: bool = True,
-        overwrite: bool = True,
-    ) -> Union[AutoShutdownFuture, Future, FileValidator, None]:
-        """
-        Start the download process.
-
-        Parameters:
-            url (str): The URL to download from.
-            file_path (str, Optional): The path to save the downloaded file. If not provided, the file is saved in the current working directory.
-                If `file_path` is a directory, the file is saved in that directory. If `file_path` is a file name, the file is saved with that name.
-            segments (int, Optional): The number of segments to divide the file into for multi-threaded download. Default is 10.
-            display (bool, Optional): Whether to display download progress and other messages. Default is True.
-            multisegment (bool, Optional): Whether to use multi-Segment download. Default is True.
-            block (bool, Optional): Whether to block the function until the download is complete. Default is True.
-            retries (int, Optional): The number of times to retry the download if it fails. Default is 0.
-            mirror_func (Callable[[], str], Optional): A function that returns a new download URL if the download fails. Default is None.
-            etag (bool, Optional): Whether to validate the ETag before resuming downloads. Default is True.
-            overwrite (bool, Optional): Whether to overwrite the file if it already exists. Default is True.
-
-        Returns:
-            AutoShutdownFuture: If `block` is False.
-            FileValidator: If `block` is True and the download successful.
-            None: If `block` is True and the download fails.
-        """
-
-        def download():
-            for i in range(retries + 1):
-                try:
-                    _url = mirror_func() if i > 0 and callable(mirror_func) else url
-                    if i > 0:
-                        logging.info("Retrying... (%d/%d)", i, retries)
-
-                    self._reset()
-                    result = self._execute(
-                        _url,
-                        file_path,
-                        segments,
-                        display,
-                        multisegment,
-                        etag,
-                        overwrite,
-                    )
-
-                    if self._stop or self.completed:
-                        if display:
-                            print(f"Time elapsed: {seconds_to_hms(self.time_spent)}")
-                        return result
-
-                    time.sleep(3)
-
-                except Exception as e:
-                    logging.error("(%s) [%s]", e.__class__.__name__, e)
-
-            self._status = 1
-            self.failed = True
-            return None
-
-        if self._allow_reuse:
-            future = self._pool.submit(download)
-        else:
-            future = AutoShutdownFuture(self._pool.submit(download), [self._pool])
-
-        if block:
-            result = future.result()
-            return result
-
-        return future
-
-    def stop(self) -> None:
-        """Stop the download process."""
-        self._interrupt.set()
-        self._stop = True
-        time.sleep(1)  # wait for threads
-
-    def shutdown(self):
-        self._pool.shutdown()
-
-
-class PypdlFactory:
+class PypdlFactory(Factory):
     """
     A factory class for managing multiple instances of the Pypdl downloader.
 
-    Keyword Arguments:
-        params (dict, optional): A dictionary, list of tuples or bytes to send as a query string. Default is None.
-        allow_redirects (bool, optional): A Boolean to enable/disable redirection. Default is True.
-        auth (tuple, optional): A tuple to enable a certain HTTP authentication. Default is None.
-        cert (str or tuple, optional): A String or Tuple specifying a cert file or key. Default is None.
-        cookies (dict, optional): A dictionary of cookies to send to the specified url. Default is None.
-        headers (dict, optional): A dictionary of HTTP headers to send to the specified url. Default is None.
-        proxies (dict, optional): A dictionary of the protocol to the proxy url. Default is None.
-        timeout (number or tuple, optional): A number, or a tuple, indicating how many seconds to wait for the client to make a connection and/or send a response. Default is 10 seconds.
-        verify (bool or str, optional): A Boolean or a String indication to verify the servers TLS certificate or not. Default is True.
+    This class accepts keyword arguments that are valid for Pypdl.
     """
-
-    def __init__(self, instances: int = 2, allow_reuse=False, **kwargs):
-        self._instances = [Pypdl(True, **kwargs) for _ in range(instances)]
-        self._allow_reuse = allow_reuse
-        self._pool = ThreadPoolExecutor(max_workers=2)
-        self._stop = False
-        self._prog = True
-        self._completed_size = 0
-        self._completed_prog = 0
-        self._lock = Event()
-        self._running = []
-
-        self.progress = 0
-        self.speed = 0
-        self.time_spent = 0
-        self.current_size = 0
-        self.total = 0
-        self.completed = []
-        self.failed = []
-        self.remaining = []
-
-    def _reset(self):
-        self._stop = False
-        self._prog = True
-        self._completed_size = 0
-        self._completed_prog = 0
-        self._running.clear()
-        self.completed.clear()
-        self.failed.clear()
-        self.remaining.clear()
-
-    def _calc_values(self):
-        def sum_attribute(instances, attribute):
-            return sum(getattr(instance, attribute) for instance in instances)
-
-        def average_attribute(instances, attribute, total):
-            return sum_attribute(instances, attribute) // total
-
-        if self._running:
-            self.speed = average_attribute(self._running, "speed", len(self._running))
-
-            self.progress = (
-                average_attribute(self._running, "progress", self.total)
-                + self._completed_prog
-            )
-            self.current_size = (
-                sum_attribute(self._running, "current_size") + self._completed_size
-            )
-
-    def _display(self):
-        cursor_up()
-        completed = len(self.completed)
-        prog = (
-            all(instance.size for instance in self._instances) if self._prog else False
-        )
-
-        if prog:
-            progress_bar = f"[{'█' * self.progress}{'·' * (100 - self.progress)}] {self.progress}% \n"
-            info = f"Total: {completed}/{self.total}, Downloaded: {to_mb(self.current_size):.2f} MB, Speed: {self.speed:.2f} MB/s "
-            print(progress_bar + info)
-        else:
-            download_stats = f"Downloading...{" " * 95}\n"
-            info = f"Total: {completed}/{self.total}, Downloaded: {to_mb(self.current_size):.2f} MB, Speed: {self.speed:.2f} MB/s "
-            print(download_stats + info)
-
-    def _compute(self, display):
-        with ScreenCleaner(display):
-            while len(self.completed) + len(self.failed) < self.total:
-                if not self._lock.is_set():
-                    self._calc_values()
-
-                    if display:
-                        self._display()
-
-                    if self._stop:
-                        break
-                time.sleep(0.5)
-
-            self.progress = self._completed_prog
-            self.current_size = self._completed_size
-            if display:
-                self._display()
-                print(f"Time elapsed: {seconds_to_hms(self.time_spent)}")
-
-    def _add_future(self, instance, task, futures):
-        url, *kwargs = task
-        instance._status = None
-        kwargs = kwargs[0] if kwargs else {}
-        kwargs.update({"block": False, "display": False, "overwrite": False})
-        future = instance.start(url, **kwargs)
-        futures[future] = (instance, url)
-        while instance._status is None:
-            time.sleep(0.1)
-
-    def _execute(self, tasks):
-        start_time = time.time()
-        self.total = len(tasks)
-        self.remaining = tasks[len(self._instances) :]
-        futures = {}
-
-        for instance, task in zip(self._instances, tasks):
-            self._add_future(instance, task, futures)
-            self._running.append(instance)
-
-        self._lock.clear()
-
-        while len(self.completed) + len(self.failed) < self.total:
-            if self._stop:
-                break
-
-            for future in as_completed(futures):
-                instance, curr_url = futures.pop(future)
-
-                if instance.completed:
-                    self._lock.set()
-                    if instance.size:
-                        self._completed_size += instance.size
-                    else:
-                        self._prog = False
-                    self._completed_prog += int((1 / self.total) * 100)
-                    self.completed.append((curr_url, future.result()))
-                elif instance.failed:
-                    self._lock.set()
-                    self.failed.append(curr_url)
-                    logging.error("Download failed: %s", curr_url)
-                else:
-                    continue
-
-                if self.remaining:
-                    self._add_future(instance, self.remaining.pop(0), futures)
-                else:
-                    self._running.remove(instance)
-
-                if len(self.completed) + len(self.failed) < self.total:
-                    self._lock.clear()
-                    time.sleep(0.5)
-
-        self.time_spent = time.time() - start_time
-        self._lock.clear()
-
-        return self.completed
-
-    def start(
-        self, tasks: list, display: bool = True, block: bool = True
-    ) -> Union[AutoShutdownFuture, Future, list]:
-        """
-        Start the download process for multiple tasks.
-
-        Parameters:
-            tasks (list): A list of tasks to be downloaded.
-                Each task is a tuple where the 1st element is the URL and the 2nd element is an optional dict with kwargs for pypdl start method.
-            display (bool, Optional): Whether to display download progress and other messages. Default is True.
-            block (bool, Optional): Whether to block the function until all downloads are complete. Default is True.
-
-        Returns:
-            AutoShutdownFuture: If `block` is False. This is a future object that can be used to check the status of the downloads.
-            list: If `block` is True. This is a list of tuples where each tuple contains the URL of the download and the result of the download.
-        """
-        self._reset()
-        if self._allow_reuse:
-            future = self._pool.submit(self._execute, tasks)
-        else:
-            future = AutoShutdownFuture(
-                self._pool.submit(self._execute, tasks), [*self._instances, self._pool]
-            )
-
-        self._lock.set()
-
-        if block:
-            self._compute(display)
-            result = future.result()
-            return result
-
-        self._pool.submit(self._compute, display)
-        return future
-
-    def stop(self):
-        """Stops all active downloads."""
-        self._lock.set()
-        self._stop = True
-        for instance in self._instances:
-            instance.stop()
-
-    def shutdown(self):
-        for instance in self._instances:
-            instance.shutdown()
-        self._pool.shutdown()
